Reset Button Improvement
Reset Button can now be pressed multiple times

+++ RitoConnector/CacheManager.cs
@@ -20,7 +20,10 @@ namespace RitoConnector
 
 		public static void resetCache()
 		{
-			Directory.Delete("./resources", true);
+			if (Directory.Exists("./resources"))
+			{
+				Directory.Delete("./resources", true);
+			}
 		}
 
 		public BitmapImage ProfileIcon(int ProfileIconID)
